+++ src/Adjustments/Contracts/Adjustment.php
@@ -44,6 +44,11 @@ interface Adjustment
      */
     public function isCredit(): bool;
 
+    /**
+     * Adjustments with value 0 are considered neutral
+     */
+    public function isNeutral(): bool;
+
     public function isIncluded(): bool;
 
     public function isLocked(): bool;