+++ src/Adjustments/Models/Adjustment.php
@@ -19,6 +19,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 use Illuminate\Support\Arr;
 use Konekt\Enum\Eloquent\CastsEnums;
+use Nette\Utils\Floats;
 use Vanilo\Adjustments\Contracts\Adjustable;
 use Vanilo\Adjustments\Contracts\Adjuster;
 use Vanilo\Adjustments\Contracts\Adjustment as AdjustmentContract;
@@ -143,6 +144,11 @@ class Adjustment extends Model implements AdjustmentContract
     public function isCredit(): bool
     {
         return $this->amount < 0;
+    }
+
+    public function isNeutral(): bool
+    {
+        return Floats::isZero($this->amount);
     }
 
     public function isIncluded(): bool